+++ dotnet/StartupExtender.cs
@@ -1,43 +1,18 @@
+using Cybersource.Data;
 using Cybersource.Services;
-using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Logging;
 
 namespace Vtex
 {
     public class StartupExtender
     {
-        public void ExtendConstructor(IConfiguration config, IWebHostEnvironment env)
-        {
-
-        }
-
         public void ExtendConfigureServices(IServiceCollection services)
         {
+            services.AddTransient<ICybersourceRepository, CybersourceRepository>();
             services.AddTransient<IVtexApiService, VtexApiService>();
             services.AddSingleton<IVtexEnvironmentVariableProvider, VtexEnvironmentVariableProvider>();
             services.AddHttpContextAccessor();
             services.AddHttpClient();
-        }
-
-        // This method is called inside Startup.Configure() before calling app.UseRouting()
-        public void ExtendConfigureBeforeRouting(IApplicationBuilder app, IWebHostEnvironment env)
-        {
-
-        }
-
-        // This method is called inside Startup.Configure() before calling app.UseEndpoint()
-        public void ExtendConfigureBeforeEndpoint(IApplicationBuilder app, IWebHostEnvironment env)
-        {
-
-        }
-
-        // This method is called inside Startup.Configure() after calling app.UseEndpoint()
-        public void ExtendConfigureAfterEndpoint(IApplicationBuilder app, IWebHostEnvironment env)
-        {
-
         }
     }
 }