+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.0.15] - 2022-08-05
+
 ### Added
 
 - (CYBRSOURCE-37) Logging optimization.